chore: clear the clippy leg of the gate, and a crash it was hiding
`ci.yml` runs clippy with `-D warnings`, and there were 18. Twelve were
mechanical and applied by `clippy --fix`, which only touches what the
tool marks as equivalent: format arguments inlined, `.map(Some)`, an
`assert_eq!` against a boolean literal.

One was this work's and is fixed as code rather than silenced: two arms
of `shape_of` returned the same shape, which said "these are different
cases" about two spellings of one.

The remaining five are in `processing/repository.rs` — persistence
functions with one argument per column, and tuples that mirror a
SELECT's column list position for position. They get `#[allow]` with the
reason written down: refactoring a 2,400-line file nobody asked about,
two days from a release, is risk without benefit, and a named tuple
would put a layer between the query and its destructuring, which is the
thing most likely to drift from the SQL.

Separately, the desktop test job was exiting 1 with 1437 tests passing:
one unhandled error is enough for Vitest. The error was real rather than
a test artifact — `snapshot.actions` arrived `undefined` and the panel
took itself down reading `.length`. The store trusted whatever the IPC
returned; it checks the shape now, because a response with an
unexpected shape should not crash a panel, and that path reached none of
the error handling that already existed.

+++ apps/desktop/src-tauri/src/processing/recovery.rs
@@ -135,7 +135,7 @@ pub fn recover_session(
                 [],
             )
             .map_err(|e| format!("Failed to close attempts: {e}"))?;
-        out.attempts_closed = closed as usize;
+        out.attempts_closed = closed;
         // Batches: running work waits for resume; confirmed intents converge.
         // `user` only. A `repair`/`manual` batch is a long-lived container
         // that is always running with an empty complete snapshot — parking one
@@ -178,7 +178,7 @@ pub fn recover_session(
             [],
         )
         .map_err(|e| format!("Failed to pause batches: {e}"))?;
-        out.batches_paused = pausing as usize;
+        out.batches_paused = pausing;
         // Unfinished cancellations converge even across the crash.
         let cancelling: Vec<String> = conn
             .prepare("SELECT id FROM processing_batches WHERE state = 'cancelling'")
